Moved Initialization logic of MaliciousSiteProtection module in MaliciousSiteProtectionService
