+++ DuckDuckGo/AppDependencies.swift
@@ -52,7 +52,7 @@ struct AppDependencies {
     let widgetRefreshModel: NetworkProtectionWidgetRefreshModel
     let autofillPixelReporter: AutofillPixelReporter
     let crashReportUploaderOnboarding: CrashCollectionOnboarding
-    let maliciousSiteProtectionManager: MaliciousSiteProtectionManaging
+    let maliciousSiteProtectionService: MaliciousSiteProtectionService
 
     var syncDidFinishCancellable: AnyCancellable?
