+++ DuckDuckGo/AppLifecycle/AppStates/Launching.swift
@@ -34,7 +34,6 @@ import Common
 import Combine
 import PixelKit
 import PixelExperimentKit
-import MaliciousSiteProtection
 
 /// Represents the transient state where the app is being prepared for user interaction after being launched by the system.
 /// - Usage:
@@ -77,10 +76,9 @@ struct Launching: AppState {
     private let isTesting = ProcessInfo().arguments.contains("testing")
     private let didFinishLaunchingStartTime = CFAbsoluteTimeGetCurrent()
     private let crashReportUploaderOnboarding: CrashCollectionOnboarding
-    private let maliciousSiteProtectionManager: MaliciousSiteProtectionManaging
-    private let maliciousSiteProtectionPreferencesManager = MaliciousSiteProtectionPreferencesManager()
 
     // These should ideally be let properties instead of force-unwrapped. However, due to various initialization paths, such as database completion blocks, setting them up in advance is currently not feasible. Refactoring will be done once this code is streamlined.
+    private let maliciousSiteProtectionService: MaliciousSiteProtectionService
     private let uiService: UIService
     private let unService: UNService
     private let syncDataProviders: SyncDataProviders
@@ -405,48 +403,7 @@ struct Launching: AppState {
         privacyProDataReporter.injectTabsModel(tabsModel)
 
         // Malicious Site Protection
-        let maliciousSiteProtectionAPI = MaliciousSiteProtectionAPI()
-
-        let maliciousSiteProtectionDataManager = MaliciousSiteProtection.DataManager(
-            fileStore: MaliciousSiteProtection.FileStore(
-                dataStoreURL: FileManager.default.urls(
-                    for: .applicationSupportDirectory,
-                    in: .userDomainMask
-                )
-                .first!
-            ),
-            embeddedDataProvider: nil,
-            fileNameProvider: MaliciousSiteProtectionManager.fileName(for:)
-        )
-
-        let maliciousSiteProtectionFeatureFlagger = MaliciousSiteProtectionFeatureFlags(featureFlagger: AppDependencyProvider.shared.featureFlagger)
-
-        let remoteIntervalProvider: (MaliciousSiteProtection.DataManager.StoredDataType) -> TimeInterval = { dataKind in
-            switch dataKind {
-            case .hashPrefixSet: .minutes(maliciousSiteProtectionFeatureFlagger.hashPrefixUpdateFrequency)
-            case .filterSet: .minutes(maliciousSiteProtectionFeatureFlagger.filterSetUpdateFrequency)
-            }
-        }
-
-        let updateManager = MaliciousSiteProtection.UpdateManager(
-            apiEnvironment: maliciousSiteProtectionAPI.environment,
-            service: maliciousSiteProtectionAPI.service,
-            dataManager: maliciousSiteProtectionDataManager,
-            updateIntervalProvider: remoteIntervalProvider
-        )
-
-        let maliciousSiteProtectionDatasetsFetcher = MaliciousSiteProtectionDatasetsFetcher(
-            updateManager: updateManager,
-            featureFlagger: maliciousSiteProtectionFeatureFlagger,
-            userPreferencesManager: maliciousSiteProtectionPreferencesManager
-        )
-
-        maliciousSiteProtectionManager = MaliciousSiteProtectionManager(
-            dataFetcher: maliciousSiteProtectionDatasetsFetcher,
-            api: maliciousSiteProtectionAPI,
-            preferencesManager: maliciousSiteProtectionPreferencesManager,
-            maliciousSiteProtectionFeatureFlagger: maliciousSiteProtectionFeatureFlagger
-        )
+        maliciousSiteProtectionService = MaliciousSiteProtectionService(featureFlagger: AppDependencyProvider.shared.featureFlagger)
 
         if shouldPresentInsufficientDiskSpaceAlertAndCrash {
             window = UIWindow(frame: UIScreen.main.bounds)
@@ -483,8 +440,8 @@ struct Launching: AppState {
                                                     textZoomCoordinator: Self.makeTextZoomCoordinator(),
                                                     websiteDataManager: Self.makeWebsiteDataManager(fireproofing: fireproofing),
                                                     appDidFinishLaunchingStartTime: didFinishLaunchingStartTime,
-                                                    maliciousSiteProtectionManager: maliciousSiteProtectionManager,
-                                                    maliciousSiteProtectionPreferencesManager: maliciousSiteProtectionPreferencesManager)
+                                                    maliciousSiteProtectionManager: maliciousSiteProtectionService.manager,
+                                                    maliciousSiteProtectionPreferencesManager: maliciousSiteProtectionService.preferencesManager)
 
             mainViewController!.loadViewIfNeeded()
             syncErrorHandler.alertPresenter = mainViewController
@@ -579,7 +536,7 @@ struct Launching: AppState {
         tipKitAppEventsHandler.appDidFinishLaunching()
 
         // Register Malicious Site Protection background tasks to fetch datasets
-        maliciousSiteProtectionManager.registerBackgroundRefreshTaskHandler()
+        maliciousSiteProtectionService.onLaunching()
     }
 
     private var appDependencies: AppDependencies {
@@ -606,7 +563,7 @@ struct Launching: AppState {
             widgetRefreshModel: widgetRefreshModel,
             autofillPixelReporter: autofillPixelReporter,
             crashReportUploaderOnboarding: crashReportUploaderOnboarding,
-            maliciousSiteProtectionManager: maliciousSiteProtectionManager
+            maliciousSiteProtectionService: maliciousSiteProtectionService
         )
     }
